video-processor/requirements.txt
flask==3.0.0
flask-cors==4.0.0
ffmpeg-python==0.2.0
openai==1.3.0
requests==2.31.0
python-dotenv==1.0.0
google-cloud-videointelligence==2.11.4
pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
